% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.r
\name{coef.eicm}
\alias{coef.eicm}
\title{Extract EICM model coefficients}
\usage{
\method{coef}{eicm}(object, ...)
}
\arguments{
\item{object}{a EICM model.}

\item{...}{additional argument(s) for methods.}
}
\value{
A list with three coefficient matrices:
  \describe{
    \item{$env}{environmental coefficients}
    \item{$sp}{species interaction coefficients. It reads as: species C (column) affects species R (row) with the coefficient sp[R, C]}
    \item{$samples}{estimated latent variable values}
  }
}
\description{
Extract the EICM model coefficients, organized in three separate matrices.
}
