% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.r
\name{isCyclic}
\alias{isCyclic}
\title{Is the adjacency matrix a cyclic graph?}
\usage{
isCyclic(coefs)
}
\arguments{
\item{coefs}{a square adjacency matrix.}
}
\value{
A logical indicating whether there are cycles in the graph.
}
\description{
Tests whether the given graph (given as an adjacency matrix) contains cycles.
}
