% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.r
\name{logLik.eicm}
\alias{logLik.eicm}
\title{EICM (penalized) log-likelihood}
\usage{
\method{logLik}{eicm}(object, occurrences = NULL, allow.na = TRUE, ...)
}
\arguments{
\item{object}{a EICM model}

\item{occurrences}{the occurrence data matrix. If omitted, the data matrix used to fit the model is used.}

\item{allow.na}{logical. Allow NAs in the occurrence matrix? If no NAs exist, it's faster to set to FALSE.}

\item{...}{additional argument(s) for methods.}
}
\value{
A \code{logLik} object.
}
\description{
Compute the (penalized) log-likelihood of the data matrix included in the EICM model,
or the log-likelihood of a new occurrence data matrix given the model.
}
