% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{plot.confint.eicm}
\alias{plot.confint.eicm}
\title{Plot EICM estimates and confidence intervals}
\usage{
\method{plot}{confint.eicm}(x, truemodel = NULL, ...)
}
\arguments{
\item{x}{a eicm.confint object}

\item{truemodel}{for validation purposes only. The true model used to simulate data.}

\item{...}{other arguments passed to other functions.}
}
\value{
NULL.
}
\description{
Plot EICM estimates and confidence intervals, in a dot-and-whisker plot.
}
