% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{plot.profile.eicm}
\alias{plot.profile.eicm}
\title{Plot EICM likelihood profile}
\usage{
plot.profile.eicm(x, level = 0.99, ...)
}
\arguments{
\item{x}{a profile.eicm object, created with \code{\link{profile.eicm}}.}

\item{level}{the significance level desired.}

\item{...}{other arguments passed to other functions.}
}
\value{
NULL.
}
\description{
Plot one likelihood profile with the line representing a confidence level threshold.
}
