% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.r
\name{print.eicm.matrix}
\alias{print.eicm.matrix}
\title{Print EICM model matrix}
\usage{
\method{print}{eicm.matrix}(x, ...)
}
\arguments{
\item{x}{a EICM model matrix.}

\item{...}{additional argument(s) for methods.}
}
\value{
NULL.
}
\description{
Prints an excerpt of the EICM model coefficients.
}
