% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{truemodel}
\alias{truemodel}
\title{A parameterized EICM model for simulation}
\format{A 'eicm' object with 2 environmental predictors in 400 samples,
32 species and 30 species interactions.}
\usage{
truemodel
}
\description{
A parameterized EICM model ready for simulating communities with a frequency
distribution following a Beta distribution with shape1=1.5 and shape2=3.
}
\details{
The model was generated with the command:

\code{truemodel <- generateEICM(32, 400, 2, 30, shape1=1.5, shape2=3)}
}
\keyword{datasets}
